placed the old tracker with the new one and updated stage1 to use the playback controller
